services/whisper/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
faster-whisper>=1.0.3
PyJWT>=2.8.0
huggingface-hub>=0.23.0
